% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-internal.r
\name{binariseMask}
\alias{binariseMask}
\title{Binarise mask.}
\usage{
binariseMask(mask, thresh = 0.9)
}
\arguments{
\item{mask}{An image of class \code{nifti}.}

\item{thresh}{Treshhold for binarisation of images and masks. A threshold near 1 (say 0.9) is conservative according to FLIRT FAQ.}
}
\value{
A binarised image.
}
\description{
Binarise mask.
}
\keyword{internal}

