% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{downsample}
\alias{downsample}
\title{Downsample an image.}
\usage{
downsample(img, subFactor = 1, method = c("FLIRT", "SUBSAMP2"),
  bin = FALSE, thresh = 0.9)
}
\arguments{
\item{img}{An image of class \code{nifti}.}

\item{subFactor}{Downsampling factor.}

\item{method}{Downsampling method.}

\item{bin}{\code{logical}. Set to \code{TRUE} for downsampling of binary images or masks.}

\item{thresh}{Treshhold for re-binarisation of downsampled binary images and masks. A threshold near 1 (say 0.9) is conservative according to FLIRT FAQ.}
}
\value{
A downsampled image.
}
\description{
Downsample an image.
}

