% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{getSNPfdr}
\alias{getSNPfdr}
\title{Get False Discovery Rates.}
\usage{
getSNPfdr(p.value, eff.no.tests)
}
\arguments{
\item{p.value}{A vector of raw p-values.}

\item{eff.no.tests}{Efficient number of performed tests.}
}
\value{
A summary of FDR q-values.
}
\description{
Get False Discovery Rates.
}

