% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{getSNPresults}
\alias{getSNPresults}
\title{Get vGWAS results for selected SNPs.}
\usage{
getSNPresults(meh, sel.snps)
}
\arguments{
\item{meh}{vGWAS results.}

\item{sel.snps}{\code{character} vector of SNP names.}
}
\value{
A named list of results for each SNP.
}
\description{
Get vGWAS results for selected SNPs.
}

