% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-internal.r
\name{readFlatROI}
\alias{readFlatROI}
\title{Read flattened ROI from a NIfTI file.}
\usage{
readFlatROI(path, subFactor = 0, voxelList = NULL, method = c("FLIRT",
  "SUBSAMP2"))
}
\arguments{
\item{path}{A path to the NIfTI file.}

\item{subFactor}{Downsampling factor.}

\item{voxelList}{List of chosen voxel coordinates. If NULL, return the whole file.}

\item{method}{Downsampling method.}
}
\value{
Vector of chosen voxel intensities.
}
\description{
Read flattened ROI from a NIfTI file.
}
\keyword{internal}

