% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{readFlatROIs}
\alias{readFlatROIs}
\title{Read flattened ROIs from a list of NIfTI files using a cluster of choice.}
\usage{
readFlatROIs(paths, ids, subFactor = 0, mask = NULL, method = c("FLIRT",
  "SUBSAMP2"), nCores = getOption("mc.cores", detectCores(logical = FALSE)),
  clType = "PSOCK", makeCluster = TRUE, userCluster = NULL)
}
\arguments{
\item{paths}{A list of paths to the NIfTI files.}

\item{ids}{A list of subject IDs.}

\item{subFactor}{Downsampling factor.}

\item{mask}{Image mask (a \code{nifti} object or an array).}

\item{method}{Downsampling method.}

\item{nCores}{number of cores to use. Default: all physical cores available}

\item{clType}{cluster type. Default: \code{PSOCK}}

\item{makeCluster}{\code{logical}. \code{TRUE} if a cluster should be made (default).}

\item{userCluster}{Cluster to use specified by the user. If \code{NULL} and makeCluster is \code{FALSE}, a default cluster will be used.}
}
\value{
Array of chosen voxel intensities for all the subjects.
}
\description{
Read flattened ROIs from a list of NIfTI files using a cluster of choice.
}

