% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{report2mni}
\alias{report2mni}
\title{Convert snp report to MNI coordinates}
\usage{
report2mni(path)
}
\arguments{
\item{path}{Path to report file.}
}
\description{
Convert snp report to MNI coordinates
}

