% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen.r
\docType{package}
\name{RImaGen}
\alias{RImaGen}
\alias{RImaGen-package}
\title{RImaGen: An R package for voxelwise genome-wide association studies.}
\description{
This package allows you to perform vGWAS analyses as described in Stein et al., 2010, but faster.
}

