% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{selectSNPs}
\alias{selectSNPs}
\title{Select SNPs by name from a \code{\linkS4class{SlicedData}} object.}
\usage{
selectSNPs(sel.snps, snpData)
}
\arguments{
\item{sel.snps}{\code{character} vector of SNP names.}

\item{snpData}{\code{\linkS4class{SlicedData}} object containing SNP data.}
}
\value{
Array containing selected SNP data.
}
\description{
Select SNPs by name from a \code{\linkS4class{SlicedData}} object.
}

