% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{toMni}
\alias{toMni}
\title{Convert voxel coordinates to MNI coordinates.}
\usage{
toMni(v, res = 2)
}
\arguments{
\item{v}{A 3D vector of voxel coordinates.}

\item{res}{Resolution in mm.}
}
\value{
MNI coordinates.
}
\description{
Convert voxel coordinates to MNI coordinates.
}

