% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{vGWAS}
\alias{vGWAS}
\title{Perform voxelwise genome-wide association analysis.}
\usage{
vGWAS(snpData, voxelData, cvrt, useModel, prescan,
  errorCovariance = numeric(), pvalue.hist = FALSE, verbose = TRUE)
}
\arguments{
\item{snpData}{\code{\linkS4class{SlicedData}} object containing selected SNPs.}

\item{voxelData}{\code{\linkS4class{SlicedData}} object containing flat voxel ROIs.}

\item{cvrt}{\code{\linkS4class{SlicedData}} objact containing covariates.}

\item{useModel}{Regression model to use.}

\item{prescan}{\code{logical}. \code{TRUE} for preliminary analysis allowing "winning SNP" ranking.}

\item{errorCovariance}{\code{numeric}. The error covariance matrix. Use \code{numeric()} for homoskedastic independent errors.}

\item{pvalue.hist}{\code{logical}, \code{numeric} or \code{"qq-plot"}.
Defines whether and how the distribution of p-values is recorded in the returned object.
Set to \code{FALSE} for a faster analysis. To record information for a histogram set to the number of bins.
To record information for a qq-plot, set to \code{qq-plot}.
Use \code{plot} on the returned object.}

\item{verbose}{\code{logical}. Set to \code{TRUE} for verbose GWAS progress reports.}
}
\value{
vGWAS results.
}
\description{
Checks for associations prepared \code{\linkS4class{SlicedData}} objects containing
flattened image ROIs, genomic data and covariates. Wrapper for \code{\link[MatrixEQTL]{Matrix_eQTL_engine}}.
}

