% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{visualiseVox}
\alias{visualiseVox}
\title{Visualise by-voxel statistics for each voxel, e.g winning SNP p-value per voxel.}
\usage{
visualiseVox(result, mask, ref.img, pv.range = NULL, log.cutoff = 0,
  plot = TRUE, title = "winning SNP", titleSize = 2,
  beforeTitle = "-log10 p-values by\\n", afterTitle = "", outPath = NULL,
  matPath = NULL, coordsOnly = FALSE)
}
\arguments{
\item{result}{Vector of target statistics.}

\item{mask}{Image mask.}

\item{ref.img}{Reference image.}

\item{pv.range}{P-value range for scale.}

\item{log.cutoff}{Negative log p-value cutoff value.}

\item{plot}{\code{logical}. Set to \code{TRUE} for automated plotting of the results.}

\item{title}{Plot title.}

\item{titleSize}{Title font size.}

\item{beforeTitle}{Text to print before the title.}

\item{afterTitle}{Text to print after the title.#'}

\item{outPath}{Output path. If \code{NULL} (default) no plot files will be saved.}

\item{matPath}{Path to convolution matrix for coregistration of results to the reference image. Set to \code{NULL} if in the same space.}

\item{coordsOnly}{\code{logical}. Set to \code{TRUE} to return coordinates of the best voxel only.}
}
\value{
A list containing the results visualisation image and a vector of coordinates for the max. value voxel.
}
\description{
Visualise by-voxel statistics for each voxel, e.g winning SNP p-value per voxel.
}

