% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-main.R
\name{performVGWAS}
\alias{performVGWAS}
\title{Perform voxelwise genome-wide association study.}
\usage{
performVGWAS(genePath, brainPath, mdtPath, maskPath, infoPath, subFactor,
  outResolution = 2^(subFactor - 1), outPath = NULL, force.snps = NULL,
  useModel = modelLINEAR, top.thresh = 5, log.cutoff = 4,
  eff.no.tests = 275575, sampleSize = NULL, mockPath = NULL,
  verbose.gwas = FALSE)
}
\arguments{
\item{genePath}{Path to a directory containing plink files with genomic data.}

\item{brainPath}{Path to a directory containing imaging data.}

\item{mdtPath}{Path to a referance image.}

\item{maskPath}{Path to an image mask.}

\item{infoPath}{Path to .csv containing subject demographics.}

\item{subFactor}{Downsampling factor.}

\item{outResolution}{Output images resolution.}

\item{outPath}{Path to output directory.}

\item{force.snps}{\code{character} vector of SNPs forced to be analysed even if not passing the quality control.}

\item{useModel}{Regression model to use.}

\item{top.thresh}{Number of top SNPs to be analysed and visualised.}

\item{log.cutoff}{Negative log p-value cutoff value for results visualisation.}

\item{eff.no.tests}{Effective number of tests (SNPs) for p-value correction.}

\item{sampleSize}{Subject sample size.}

\item{mockPath}{Path to input directory with saved objects for vGWAS mocking. Mock parameters should match  original parameters. \code{NULL} for real analyses.}

\item{verbose.gwas}{\code{logical}. Set to \code{TRUE} for verbose GWAS progress reports.}
}
\description{
Loads and prepares the data, runs preliminary (if necessary) and full analyses.
Then visualises the results.
}

