% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-main.R
\name{performVGWAS}
\alias{performVGWAS}
\title{Perform voxelwise genome-wide association study.}
\usage{
performVGWAS(genePath, niiFiles, niiIDs, ref.imgPath, maskPath, infoPath,
  subFactor, covar = character(), errorCovariance = numeric(),
  out.subFactor = subFactor, matPath = NULL, outPath = NULL,
  force.snps = NULL, useModel = MatrixEQTL::modelLINEAR, top.thresh = 5,
  log.cutoff = 4, eff.no.tests = 275575, sampleSize = NULL,
  randomSample = FALSE, mockPath = NULL)
}
\arguments{
\item{genePath}{Path to a directory containing plink files with genomic data.}

\item{niiFiles}{Paths to files containing imaging data.}

\item{niiIDs}{Subject IDs for each image. Shuffle for permutation tests.}

\item{ref.imgPath}{Path to a referance image.}

\item{maskPath}{Path to an image mask.}

\item{infoPath}{Path to .csv containing subject demographics.}

\item{subFactor}{Downsampling factor.}

\item{covar}{Covariates matrix with subject IDs as column names.}

\item{errorCovariance}{Covariance matrix for the error term. Set to \code{numeric()} for multiple of identity (default, most cases).}

\item{out.subFactor}{Output images downsampling factor. Default: \code{max(1, subFactor-1)}.}

\item{matPath}{Path to convolution matrix for coregistration of results to the reference image. Set to \code{NULL} if in the same space.}

\item{outPath}{Path to output directory.}

\item{force.snps}{\code{character} vector of SNPs forced to be analysed even if not passing the quality control.}

\item{useModel}{Regression model to use.}

\item{top.thresh}{Number of top SNPs to be analysed and visualised.}

\item{log.cutoff}{Negative log p-value cutoff value for results visualisation.}

\item{eff.no.tests}{Effective number of tests (SNPs) for p-value correction.}

\item{sampleSize}{Subject sample size.}

\item{randomSample}{\code{logical}. Set to \code{TRUE} for subject sample randomisation.}

\item{mockPath}{Path to input directory with saved objects for vGWAS mocking. Mock parameters should match  original parameters.
Useful for registering results to different standard spaces and resolutions, selecting alternative SNPs for the full analysys, of using different genetic models.
\code{NULL} for real analyses.}
}
\value{
A list containing all the resulting statistical parametric maps.
}
\description{
Loads and prepares the data, runs preliminary (if necessary) and full analyses.
Then visualises the results.
}

