% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImaGen-functions.r
\name{visualiseSNP}
\alias{visualiseSNP}
\title{Visualise by-voxel statistics for a chosen SNP, e.g that SNP's p-value per voxel.}
\usage{
visualiseSNP(snp.name, results, mask, ref.img, log.cutoff = 0, plot = TRUE,
  title = snp.name, titleSize = 1, beforeTitle = "-log10 p-values by\\n",
  afterTitle = "", outPath = NULL, matPath = NULL)
}
\arguments{
\item{snp.name}{Chosen SNP name.}

\item{results}{A named list of results for each SNP.}

\item{mask}{Image mask (a \code{nifti} object or an array).}

\item{ref.img}{Reference image.}

\item{log.cutoff}{Negative log p-value cutoff value.}

\item{plot}{\code{logical}. Set to \code{TRUE} for automated plotting of the results.}

\item{title}{Plot title.}

\item{titleSize}{Title font size.}

\item{beforeTitle}{Text to print before the title.}

\item{afterTitle}{Text to print after the title.}

\item{outPath}{Output path. If \code{NULL} (default) no plot files will be saved.}

\item{matPath}{Path to convolution matrix for coregistration of results to the reference image. Set to \code{NULL} if in the same space.}
}
\value{
Results visualisation image.
}
\description{
Visualise by-voxel statistics for a chosen SNP, e.g that SNP's p-value per voxel.
}

